comment out upgrade test as it doesnt do what we need
